+++ src/App.js
@@ -1,16 +1,12 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import Login from './pages/Login';
-import Wallet from './pages/Wallet';
 
 function App() {
   return (
     <Switch>
       <Route exact path="/">
         <Login />
-      </Route>
-      <Route exact path="/carteira">
-        <Wallet />
       </Route>
     </Switch>
   );
