+++ src/actions/index.js
@@ -1,8 +1,27 @@
 // Coloque aqui suas actions
-import userActions from './userActions';
-import walletActions from './walletActions';
+import ACTIONS_TYPES from '../constants';
+import services from '../services';
+
+export const authenticate = (credentials) => async (dispatch) => {
+  dispatch({
+    type: ACTIONS_TYPES.LOGIN_REQUEST,
+  });
+
+  try {
+    const user = await services.login(credentials);
+
+    dispatch({
+      type: ACTIONS_TYPES.LOGIN_SUCCESS,
+      payload: { ...user },
+    });
+  } catch (error) {
+    dispatch({
+      type: ACTIONS_TYPES.LOGIN_ERROR,
+      payload: error,
+    });
+  }
+};
 
 export default {
-  ...userActions,
-  ...walletActions,
+  authenticate,
 };
