--- src/components/ExpenseForm/styles.css
@@ -1,26 +1,0 @@
-.expense-form {
-  display: grid;
-  gap: 8px;
-  grid-template-columns: repeat(6, auto);
-}
-
-@media ( max-width : 1024px ) {
-
-  .expense-form {
-    grid-template-columns: repeat(3, auto);
-  }
-}
-
-@media ( max-width : 768px ) {
-
-  .expense-form {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media ( max-width : 425px ) {
-
-  .expense-form {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
