--- src/components/TableExpenses/styles.css
@@ -1,55 +1,0 @@
-.before-currency_BRL:before {
-  content: 'R$';
-}
-
-.before-code_USD:before {
-  content: 'USD';
-}
-
-.before-code_CAD:before {
-  content: 'CAD';
-}
-
-.before-code_EUR:before {
-  content: 'EUR';
-}
-
-.before-code_GBP:before {
-  content: 'GBP';
-}
-
-.before-code_ARS:before {
-  content: 'ARS';
-}
-
-.before-code_BTC:before {
-  content: 'BTC';
-}
-
-.before-code_LTC:before {
-  content: 'LTC';
-}
-
-.before-code_JPY:before {
-  content: 'JPY';
-}
-
-.before-code_CHF:before {
-  content: 'CHF';
-}
-
-.before-code_CNY:before {
-  content: 'CNY';
-}
-
-.before-code_ILS:before {
-  content: 'ILS';
-}
-
-.before-code_ETH:before {
-  content: 'ETH';
-}
-
-.before-code_XRP:before {
-  content: 'XRP';
-}
