+++ src/constants/walletActionsTypes.js
@@ -1,14 +1,7 @@
 const WALLET_ACTIONS_TYPES = {
-  WALLET_IS_FETCHING: 'WALLET_IS_FETCHING',
-  WALLET_ERROR: 'WALLET_ERROR',
+  LOAD_WALLET: 'LOAD_WALLET',
   LOAD_WALLET_SUCCESS: 'LOAD_WALLET_SUCCESS',
-  FETCH_CURRENCIES_SUCCESS: 'FETCH_CURRENCIES_SUCCESS',
-  ADD_EXPENSE_SUCCESS: 'ADD_EXPENSE_SUCCESS',
-  EDIT_EXPENSE_SUCCESS: 'ADD_EXPENSE_SUCCESS',
-  UPDATE_TOTAL_EXPENSES: 'UPDATE_TOTAL_EXPENSES',
-  DELETE_EXPENSE: 'DELETE_EXPENSE',
-  LOAD_RECORD_TO_EDIT: 'LOAD_RECORD_TO_EDIT',
-  SAVE_EDITED_RECORD: 'SAVE_EDITED_RECORD',
+  LOAD_WALLET_ERROR: 'LOAD_WALLET_ERROR',
 };
 
 export default WALLET_ACTIONS_TYPES;
