+++ src/pages/Login/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import useLoginPage from './useLoginPage';
-import Input from '../../components/Input';
 import './styles.css';
 
 const Login = () => {
@@ -14,7 +13,7 @@ const Login = () => {
           alt="logo trybe completa"
           className="login-container-form__brand-logo"
         />
-        <Input
+        <input
           type="email"
           className="form-control"
           id="email"
@@ -24,7 +23,7 @@ const Login = () => {
           value={ credentials.email }
           onChange={ handleChange }
         />
-        <Input
+        <input
           type="password"
           className="form-control"
           id="password"
