+++ src/pages/Wallet.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Wallet extends React.Component {
+  render() {
+    return <div>TrybeWallet</div>;
+  }
+}
+
+export default Wallet;
